refactor(domains): factory, util, repository 패키지에 있던 클래스 3개 domain 패키지로 이동
